Twitch sub helpers for frontend (#25)
# What does this change do?
Implement some helper functions in order to listen on the Twitch API for new subscriptions. This is strictly business logic for now, 

# Why is this change needed?
The overlay for the Twitch live stream (the one that shows the sponsor boxes, guest names etc) should display whenever someone subscribes, resubscribes or gifts a sub. Basically messages like:

"327jwolf just subscribed!"
"327jwolf subscribed for 4 months!"
"327jwolf gifted a subscription to CinnamonWoman"

In order for the React GUI to do this, we need to listen to the correct endpoints and figure out what data to pull out from the events.